web/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
